Put generated code in 'api' module
...instead of copying it to the consumer module's `build` directory.
Thanks to this, the `api` module is self-contained in a sense that
it can be used by the consumer as a regular dependency, without copying
files behind the scenes.

A small regression with this change: there's something wrong with task
dependencies and one has to run `./gradlew build` three times in order
for the whole app to be generated and built correctly.